refactor: qc-application workspace 도메인 패키지 정리
