+++ qc-application/src/main/java/com/eager/questioncloud/application/api/workspace/dto/WorkSpaceControllerResponse.java
@@ -1,4 +1,4 @@
-package com.eager.questioncloud.application.api.workspace;
+package com.eager.questioncloud.application.api.workspace.dto;
 
 import com.eager.questioncloud.core.domain.creator.model.CreatorProfile;
 import com.eager.questioncloud.core.domain.question.model.QuestionContent;
