+++ qc-application/src/main/java/com/eager/questioncloud/application/api/workspace/service/CreatorProfileService.java
@@ -1,4 +1,4 @@
-package com.eager.questioncloud.application.api.workspace;
+package com.eager.questioncloud.application.api.workspace.service;
 
 import com.eager.questioncloud.core.domain.creator.infrastructure.CreatorRepository;
 import com.eager.questioncloud.core.domain.creator.model.Creator;
